Rewrote the 'waitForComponentShould' step in reponse to #6
The component now takes a number of additional edge cases into
account.

disappear - element off the page / element hidden
appear - element on the page and visible

Added a range of additional unit and integration tests.

Ammended the examples to include a pause before manipulating
the dom. The examples now reflect the functionality.