Update UIBusinessSelectorContext.php
add new selector and step. 

We need this library to also be able just to look for text on a page. So domain 1 is in language foo and domain 2 is in language bar but the both have the business need to have "Disclaimer"

So the user can make a token for domain 1
Disclaimer: "Use site at own risk"

And the for domain 2
Disclaimer: "malja' Daq DeSDu' bISuDqu' 'e' lach'eghDI'"